Crate kemuler

kemuler

this crate is currently in a rapid development phase. compilation failure will probably be introduced with every 5 commits lol Update: Current API design is quite stable. But, still, it’s in development :P.

Welcome to kemuler!

This crate offers a different kind of high level input simulator framework, utilizing Rust’s type system to its full potential (or just over-engineered) instead of the usual key_down(Key).

Why? Combinators!

use kemuler::prelude::*;
// use a simulator your heart desired
// (your heart must also implement one if there's none)
let mut s = Simulator::new();

// tuple supports
(
    Key::Shift.down(),

    Char('o').click(),
    // down and then up is the same as `.click`
    Char('h').down(),
    Char('h').up(),

    Char('y').click(),
    Char('e').click(),
    Char('a').click(),
    // repeat anything before the `.repeat` 5 times,
    Char('h').click().repeat(5),

    Key::Shift.up(),
)
    // `.seq` wraps the tuple with `Sequence` type to mark it as simulatable
    .seq()
    .run_with(&mut s)
// this typed the message: "OH YEAHHHHH"

Currently, there are only a few amount of combinators present. If you’ve got some more useful combinator, please submit an issue on GitHub!

See crate’s document for more examples. Documentation currently live at https://kemuler.github.io/kemuler/kemuler.

Installation

Using Cargo command:

cargo add --git "https://github.com/kemuler/kemuler"

or add the following to your Cargo.toml dependencies:

kemuler = { git = "https://github.com/kemuler/kemuler" }

Adding kemuler on its own wouldn’t do anything. You must import a simulator (means backend) or implements one to actually simulate an input. Head to Simulators section to see cooked and ready simulator to be added to your dependencies.

Assuming using kemuler-enigo since it’s cross-platform.

Using Cargo command:

cargo add --git "https://github.com/kemuler/kemuler-enigo"

or add the following to your Cargo.toml dependencies:

kemuler-enigo = { git = "https://github.com/kemuler/kemuler-enigo" }

Simulators

Simulators that are being maintained by this crate.

kemuler-enigo

kemuler-windows

Examples

The basic

use kemuler::prelude::*;

let mut s = Simulator::new(); // use your preferred simulator

Key::Shift.down().run_with(&mut s);
Key::Shift.up().run_with(&mut s);
Char('w').down().run_with(&mut s);
Char('w').up().run_with(&mut s);

// Inputs that has been passed to the simulator
// (A world without combinator):
// Key::Shift.down();
// Key::Shift.up();
// Char('w').down();
// Char('w').up();

Combinators Overview

.then method chains simulatable together to run one by one.

use kemuler::prelude::*;

// use your preferred simul—ahh you got it.
let mut s = Simulator::new();

Key::Alt.down()
    .then(Key::Tab.down())
    .then(Key::Alt.up())
    .then(Key::Tab.up())
    .run_with(&mut s);

// Inputs that has been passed to the simulator
// (A world without combinator):
// Key::Alt.down();
// Key::Tab.down();
// Key::Alt.up();
// Key::Tab.up();

Tuple supports! Only up to 32 indexes; if you some how need much more than that then nested tuple will suffice. Chains simulatable one by one without .then. (then actually returns this)

(
    Key::Control.down(),
    MouseButton::Right.down(),
    Sleep::for_millis(100),
    Key::Control.up(),
    MouseButton::Right.up(),
)
    // wrap this tuple with the `Sequence` type to make it simulatable
    .seq()
    .run_with(&mut s);

// Inputs that has been passed to the simulator
// (A world without combinator):
// Key::Control.down();
// MouseButton::Right.down();
// Key::Control.up();
// MouseButton::Right.up();

Iterator supports! This does a for loop internally and simulate each item as long as the item is simulatable.

[
    Key::Alt.down(),
    Key::Tab.down(),
    Key::Alt.up(),
    Key::Tab.up()
]
    // wrap this iterator with the `IterSequence` type
    // to make this iterator simulatable.
    // This supports every type that implements `IntoIterator`.
    .iter_seq()
    .run_with(&mut s);

// Inputs that has been passed to the simulator
// (A world without combinator):
// Key::Alt.down();
// Key::Tab.down();
// Key::Alt.up();
// Key::Tab.up();

Other useful combinators!

// do these 3 times:
//   left click
//   space bar click
//   wait 10 millisecond
//
(
    MouseButton::Left.click(),
    Key::Space.click().sleep_for_millis(10),
)
    .seq()
    .repeat(2)
    .run_with(&mut s);

// Inputs that has been passed to the simulator
// (A world without combinator):
// MouseButton::Left.down();
// MouseButton::Left.up();
// Key::Space.down();
// Key::Space.up();
// MouseButton::Left.down();
// MouseButton::Left.up();
// Key::Space.down();
// Key::Space.up();

Modules

combinator

Module of helper combinators

common_inputs

Generally found/a standard input

input_event

Module containing base input events.

prelude

Re-exports

simulatable

Module for implementing a thing can be simulated

simulator

Module for implementing a simulator

Macros

button_like_impl_body

This macro generate convenient methods usable for traits and impl blocks. Must only be used inside trait definitons and impl block definitions.