Trait kemuler::combinator::Combine

pub trait Combine: Sized {
    // Provided methods
    fn then<S>(self, next: S) -> Sequence<(Self, S)> { ... }
    fn sleep(self, duration: Duration) -> Sequence<(Self, Sleep)> { ... }
    fn sleep_for_millis(self, duration: u64) -> Sequence<(Self, Sleep)> { ... }
    fn repeat(self, times: usize) -> Repeat<Self> { ... }
    fn iter_seq(self) -> IterSequence<Self>
       where Self: IntoIterator { ... }
    fn seq(self) -> Sequence<Self> { ... }
    fn during<DS>(self, during: DS) -> During<DS, Self>
       where DS: Invert + Clone { ... }
}

Helper combinator trait.

Provided Methods

fn then<S>(self, next: S) -> Sequence<(Self, S)>

Simulate self and then next

fn sleep(self, duration: Duration) -> Sequence<(Self, Sleep)>

Simulate self and then sleep for duration

fn sleep_for_millis(self, duration: u64) -> Sequence<(Self, Sleep)>

Simulate self and then sleep for duration in milliseconds

fn repeat(self, times: usize) -> Repeat<Self>

Repeat simulation for amount of times

fn iter_seq(self) -> IterSequence<Self>where Self: IntoIterator,

Iterate through an iterator and simulate each item Self must be an iterator and its item must be Simulatable.

fn seq(self) -> Sequence<Self>

Simulate through a tuple starting from .0. Self must be a tuple with 0 <= size <= 32. Nest them if you ever need more.

fn during<DS>(self, during: DS) -> During<DS, Self>where DS: Invert + Clone,

Simulate self during an event. After self is simulated, the during event is inverted and simulated at the end.