Trait kemuler::input_event::Invert

pub trait Invert {
    type Output;

    // Required method
    fn invert(self) -> Self::Output;
}

Trait for an event that is the inverse version of self. For example, key down is the inverse version of key up and vice versa.

The inverse must completely annihilated each other as if a + b = 0 where b is the inverse of a.

This is only implemented on SetTo<I, bool> for now since that’s the only thing that is make sense to be implemented.

Required Associated Types

type Output

Required Methods

fn invert(self) -> Self::Output