Trait kemuler::simulatable::Simulatable

pub trait Simulatable<Smlt>: Sized {
    // Required method
    fn run_with(self, simulator: &mut Smlt);
}

Simulatable is a thing that can be simulated by a simulator. In this crate, it is implemented on combinators and input event.

Required Methods

fn run_with(self, simulator: &mut Smlt)

Simulate this input.